+++ tests/typed_data_obj.rs
@@ -1,6 +1,6 @@
 use magnus::{
-    class, define_class, embed::init, eval, function, gc, method, prelude::*, ruby_handle::RubyHandle,
-    typed_data::Obj, value::Opaque, DataTypeFunctions, TypedData,
+    class, define_class, embed::init, eval, function, gc, method, prelude::*,
+    ruby_handle::RubyHandle, typed_data::Obj, value::Opaque, DataTypeFunctions, TypedData,
 };
 
 #[magnus::wrap(class = "Point", free_immediatly)]
@@ -18,7 +18,9 @@ impl Point {
 #[derive(TypedData)]
 #[magnus(class = "Line", free_immediatly, mark)]
 struct Line {
+    #[magnus(opaque_attr_reader)]
     start: Opaque<Obj<Point>>,
+    #[magnus(opaque_attr_reader)]
     end: Opaque<Obj<Point>>,
 }
 
@@ -28,16 +30,6 @@ impl Line {
             start: start.into(),
             end: end.into(),
         }
-    }
-
-    fn start(&self) -> Obj<Point> {
-        let handle = unsafe { RubyHandle::get_unchecked() };
-        handle.unwrap_opaque(self.start)
-    }
-
-    fn end(&self) -> Obj<Point> {
-        let handle = unsafe { RubyHandle::get_unchecked() };
-        handle.unwrap_opaque(self.end)
     }
 
     fn length(&self) -> f64 {